#1 [refactor] : MyScreen 컴포넌트 화
